python-listener/requirements.txt
./generated
flake8==7.1.1
requests==2.32.3
requests-sse==0.3.2
pynose==1.5.3
